program that multiplies two positive numbers.

+++ 0x0C-more_malloc_free/101-mul.c
@@ -9,7 +9,7 @@
  */
 int main(int argc, char *argv[])
 {
-	unsigned int mul;
+	unsigned long mul;
 	int i, j;
 
 	if (argc != 3)
